Move Drop-in v5 tests
