+++ tests/checkout/v5/klarna-paynow.spec.js
